Rename SectionClassifiedHeroesBlocks to SectionHeroesBlocks
Updated file and directory names to SectionHeroesBlocks for clarification. Modified imports across multiple files to reflect the new naming conventions.
