+++ frontend-next-migration/src/widgets/SectionHeroesBlocks/index.ts
