+++ frontend-next-migration/src/widgets/SectionHeroesBlocks/ui/heroesBlocks/HeroesBlocks.tsx
