+++ frontend-next-migration/src/widgets/SectionHeroesBlocks/ui/index.tsx
