+++ frontend-next-migration/src/widgets/SectionHeroesBlocks/ui/main.module.scss
